requirements.txt
pymysql
pika>=0.9.5
redis
DBUtils
chardet
requests
urllib3
psutil
arrow
parsel
aiohttp


fabric
xlwt
pytesseract
pillow
xlrd
flask
uwsgi
celery
pdfminer3k
bs4
newspaper3k
docx
win32com
pypiwin32
pyhton-docx
libxml2
libxslt
libxslt-devel
gpcharts
future
aiopg
sqlalchemy
captcha
numpy
matplotlib
opencv-python
scikit-image
scipy